Place next/font transform behind cargo feature
