+++ crates/turbopack-tests/Cargo.toml
@@ -14,8 +14,8 @@ turbopack = { path = "../turbopack" }
 
 [dev-dependencies]
 anyhow = "1.0.47"
+next-core = { path = "../next-core", features = ["next-font"] }
 once_cell = "1.13.0"
-next-core = { path = "../next-core" }
 serde = "1.0.136"
 serde_json = "1.0.85"
 similar = "2.2.0"
